+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next Version
+### Changes
+* Instead of doing a `AssetDatabase.Refresh()` to import the optimized shaders it now uses `ImportAsset()` inside a `StartAssetEditing()` and `StopAssetEditing()` block. This saves a tiny bit of time.
+
 ## v1.8
 ### Features
 * Add info text if any normal maps are found that are not using BC5 compression.